pyproject.toml
[tool.poetry]
name = "omegaconf-cloud-resolvers"
version = "0.1.0"
description = "Omegaconf custom resolvers to retrieve configuration values from cloud services"
authors = ["M Gil Valverde <m.gilvalverde@gmail.com>"]
readme = "README.md"
#packages = [{include = "hydra/"}]

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.3.0"
jmespath = "^1.0.1"


[tool.poetry.group.aws.dependencies]
boto3 = "^1.34.16"

[tool.poetry.group.gcp.dependencies]
google-cloud-secret-manager = "^2.20.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
exclude = ["_static", "build", "examples", "notebooks", "venv","_exp", "config", "dist", ".env", ".venv"]


